refactor(cli): move foundry utils to foundry.ts, move execLog to own util
