+++ packages/cli/src/commands/deprecated/call-system.ts
@@ -2,7 +2,7 @@ import { defaultAbiCoder as abi } from "ethers/lib/utils.js";
 import path from "path";
 import type { CommandModule } from "yargs";
 import { execLog } from "../../utils/deprecated/index.js";
-import { getTestDirectory } from "../../utils/forgeConfig.js";
+import { getTestDirectory } from "../../utils/foundry.js";
 
 type Options = {
   rpc?: string;
