+++ packages/cli/src/commands/deprecated/deploy-contracts.ts
@@ -2,7 +2,7 @@ import type { CommandModule } from "yargs";
 import { DeployOptions, generateAndDeploy, hsr } from "../../utils/deprecated/index.js";
 import openurl from "openurl";
 import chalk from "chalk";
-import { getSrcDirectory } from "../../utils/forgeConfig.js";
+import { getSrcDirectory } from "../../utils/foundry.js";
 
 type Options = DeployOptions & {
   watch?: boolean;
