+++ packages/cli/src/commands/deprecated/trace.ts
@@ -4,7 +4,7 @@ import { readFileSync } from "fs";
 import { Contract } from "ethers";
 import { JsonRpcProvider } from "@ethersproject/providers";
 import WorldAbi from "@latticexyz/solecs/abi/World.json" assert { type: "json" };
-import { getSrcDirectory } from "../../utils/forgeConfig.js";
+import { getSrcDirectory } from "../../utils/foundry.js";
 import path from "path";
 import { componentsDir, systemsDir } from "../../utils/deprecated/constants.js";
