+++ packages/cli/src/commands/tablegen.ts
@@ -3,7 +3,7 @@ import { mkdirSync, writeFileSync } from "fs";
 import path from "path";
 import { loadStoreConfig } from "../config/loadStoreConfig.js";
 import { renderTablesFromConfig } from "../render-table/renderTablesFromConfig.js";
-import { getSrcDirectory } from "../utils/forgeConfig.js";
+import { getSrcDirectory } from "../utils/foundry.js";
 import { formatSolidity } from "../utils/format.js";
 
 type Options = {
