+++ packages/cli/src/utils/deploy-v2.ts
@@ -2,8 +2,7 @@ import { existsSync, readFileSync, writeFileSync } from "fs";
 import path from "path";
 import { MUDConfig } from "../config/index.js";
 import { MUDError } from "./errors.js";
-import { cast, forge } from "./foundry.js";
-import { getOutDirectory, getScriptDirectory } from "./forgeConfig.js";
+import { getOutDirectory, getScriptDirectory, cast, forge } from "./foundry.js";
 import { ContractInterface, ethers } from "ethers";
 import { World } from "@latticexyz/world/types/ethers-contracts/World.js";
 import { abi as WorldABI } from "@latticexyz/world/abi/World.json";
