+++ packages/cli/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./errors.js";
-export * from "./forgeConfig.js";
+export * from "./foundry.js";
 export * from "./format.js";
 export * from "./deploy-v2.js";
-export * from "./foundry.js";
